requirements.txt
loguru

uvicorn
fastapi
starlette
sqlalchemy
psycopg2-binary
asyncpg
alembic

pydantic-settings
pydantic-extra-types
pydantic[email]
phonenumbers

contextvars
passlib
pyjwt

langgraph
langchain-core
langchain-openai
langchain-community

binance-connector
